Move the context cache hack
Moved so that non model-factory implementations can use the request
cache and included the type name in the cache key in case you are
returning a base type and inherited type within the same request.

I also shifted the type conversion code into it's own method as it would
have been a nightmare to read within the function.