Adding built in TypeConverters
This adds the following converters:

- HtmlStringConverter
- ContentPickerConverter
- MediaPickerConverter
- MultipleMediaPickerConverter
- MultiNodeTreePickerConverter

Additionally a helper class has been added (internally for now) that
aids with the conversion process by supplying a per-request-cached
UmbracoHelper instance.

I've also made HtmlString conversion part of the default conversion
pipeline. It makes more sense to me to have this run automatically than
to add an attribute to the property each time.