Update type converters to treat emtpy string the same as null
Empty strings pass through to the base TypeConverter and for enumerable
types at least, this results in a NotSupportedException.

cc90af9 added support for accepting null values, and in pretty much
every case, empty string should be treated as null: no meaninful data
can be gleaned from it for conversion.

EnumConverter works without the check because the empty string gets
pased to the base ConvertFrom method which just returns the default enum
value. But we might as well be explicit about the expected return result
for empty strings.

Fixes #58