Testing IPublishedContent with IsEnumerableOfType extension method
Whilst experimenting with returning collections of a custom `IPublishedContent` class, Ditto wouldn't attempt to cast/map them any further.

I set up a quick unit-test to see if this was the case...

```
typeof(IEnumerable<Mocks.PublishedContentMock>).IsEnumerableOfType(typeof(IPublishedContent))
```

Should this be true or false?

---

I've added in a fix that swaps around the params for the `IsAssignableFrom` check.