Unit-test UmbracoPicker processor
Following a discussion with @mattbrailsford earlier today, I started to play with mocking the UmbracoContext in order to test the UmbracoPicker processor.

So far I have manually mocked things like HttpContextBase and the PublishedContent/MediaCache objects - we may find that a mocking library (such as Moq) would be better, but I have limited knowledge on that... (totally open to advice on that!)

The interesting part about testing UmbracoPicker is that we can remove the subsequent `.As(type)` calls, as when a processor returns an `IPublishedContent` object, it will automagically check the target type and process further.