UmbracoDictionary processor - added factory
Following a discussion with @mattbrailsford earlier, (yup same chat that sparked PR #161), we agreed that using a ProcessorContext to abstract Umbraco logic wasn't the correct approach.
However the reason I used it was so I could unit-test the `UmbracoDictionary` processor.

Matt suggested that I try using a factory pattern to set the dictionary retrieval method. So this PR is my attempt.

Currently I've placed all the internal classes within the "UmbracoDictionaryAttribute.cs", I'm not 100% happy with this, but I wasn't sure where else to put them? (I'm open to suggestions)