Added blank constructor to DittoMultiProcessorAttribute
At Code Cabin, @mattbrailsford asked to be reminded to add this in so DittoMultiProcessorAttribute's base.Attributes can be populated in the Contructor body.

The benefits to this are an attribute could be populated via an optional property in the constructor body without it being required and having to be passed in to base intiialisation.