EnumerableConverter fails with inherited class
An issue has been flagged for when attempting to map a property type that inherits from a generic enumerable.

Background on the issue can be found here:
<https://our.umbraco.org/projects/developer-tools/ditto/ditto-feedback/79244-error-on-using-relatedlink-with-ditto-and-jeavon-property-converter#comment-257744>

The culprit is here: EnumerableConverterAttribute.cs, line 76
As the ojbect-type is detected as an enumerable type (line 40), but when the `GenericTypeArguments` are accessed (line 76),
the top-level type is accessed, not the base type.

These unit-tests attempt to test the processor directly, (as opposed to going via the `.As<T>` method.
The first test is successful, as it is working with a generic enumerable type directly.
The second test fails, as it is working with a class inherited from a generic enumerable type.