Crusade against UmbracoContext, DictionaryHelper removed (see desc)
Removed the static class, UmbracoDictionaryHelper.

Tentatively added in UmbracoHelper instance as in Umbraco v7.3.2 this is the best way to get a dictionary value.

Also worked in some teardowns for Resolution.Reset() so tests work between this and the UmbracoPicker tests.

Finally, there's a slight change to the dictionary tests as UmbracoHelper.GetDictionaryValue doesn't do an ordinal key check for keys. Not sure how we can fix this within Ditto, it might be a feature Umbraco adds in later on.