Attributed Type Resolver
Adds a resolver class to provide a lookup for specific Ditto attributes for a given type.
The thinking behind this is that we have a few places where we check for a custom attribute, e.g. the `UmbracoProperties` attribute inside the `UmbracoProperty` processor.
Given that this is called for each object's property, it has the potential of being queried a lot of times.

Of course this adds time to app startup, but then it's a once-off operation.