Changed how `IsDebuggingEnabled` is set.
By default the debug mode is set with the assembly compilation.
The app-startup event (e.g. when ran in an Umbraco website) will check for the `GetDebugFlag()` method.
(Same with the unit-tests)

The `IsProfilingEnabled()` method is also tied to the `IsDebuggingEnabled` flag - as Umbraco only enables profiling in debug mode.