Disable MiniProfiler feature
This is a controversial decision, it turns out that calling the Disposable Timer many (many) times is quite a performance bottleneck.
Without the profiling code the page load time reduced by 3x, (in my tests, from 250ms to 70ms).

Rather than removing the code, we have used C# compiler preprocessor directives to only include it when compiled in Debug configuration.